requirements.txt
Flask==3.0.0
Flask-SQLAlchemy=3.1.1
flask-login=0.6.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2